refactor: introduce concern-separation for ResponseBuffer, move utility network functions to separate ts file, rename parseResponseBody to readResponseText for clarity and consistency
